VehiclesRework: Use category icon as surrogate for vehicle icon
Vehicles don't have icons data provided in the configs.

It seems like we were supposed to be able to editing vehicle specific respawn configs to have an icon, see the below

```
	class vn_b_air_ch47_01_01 {
		tags[] = {"ch47","armed","transport","helicopter","army"};
        icon = VEHICLE_ICON_CAR;
	};
```

However

- this config ^ doesn't show the icon (it doesn't work)
- it would be tedious when we've already done the same for categories
- vehicle icons will likely just be the same as the category anyway

Instead, we can keep life simple, and just grab the top level category icon as a surrogate icon for each vehicle.